refactor : repacking cache config
